Finishing up back-end for category
This adds the Create/Update/Delete operations for category and a bunch of associated stored procedures and repository functions.

+++ server/src/api/models/request/category/update_category.rs
@@ -5,5 +5,5 @@ pub struct UpdateCategoryRequest {
 	pub name: String,
 
 	#[serde(rename = "parentCategoryId")]
-	pub parent_category_id: Option<i32>,
+	pub parent_category_id: Option<i64>,
 }
